[refactor]: users dto 패키지 정리(#73)
