+++ app/src/main/java/com/texthip/thip/ui/feed/screen/MySubscriptionListScreen.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.zIndex
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import com.texthip.thip.R
-import com.texthip.thip.data.model.users.FollowingList
+import com.texthip.thip.data.model.users.response.FollowingList
 import com.texthip.thip.ui.common.header.AuthorHeader
 import com.texthip.thip.ui.common.modal.ToastWithDate
 import com.texthip.thip.ui.common.topappbar.DefaultTopAppBar
